issue #1227 refactor bulkdata jbatch job codes
