+++ fhir-bulkimportexport-webapp/src/main/java/META-INF/batch-jobs/FhirBulkExportGroupChunkJob.xml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <job xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns="http://xmlns.jcp.org/xml/ns/javaee" xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee http://xmlns.jcp.org/xml/ns/javaee/jobXML_1_0.xsd" id="bulkgroupexportchunkjob" restartable="true" version="1.0">
     <listeners>
-        <listener ref="com.ibm.fhir.bulkexport.system.ExportJobListener"/>
+        <listener ref="com.ibm.fhir.jbatch.bulkdata.export.system.ExportJobListener"/>
     </listeners>
     <step id="step1">
         <chunk checkpoint-policy="item" item-count="1">
-            <reader ref="com.ibm.fhir.bulkexport.group.ChunkReader">
+            <reader ref="com.ibm.fhir.jbatch.bulkdata.export.group.ChunkReader">
                 <properties >
                     <property name="fhir.tenant" value="#{jobParameters['fhir.tenant']}"/>
                     <property name="fhir.datastoreid" value="#{jobParameters['fhir.datastoreid']}"/>                   
@@ -17,7 +17,7 @@
                     <property name="fhir.typeFilters" value="#{jobParameters['fhir.typeFilters']}"/>
                 </properties>
             </reader>
-            <writer ref="com.ibm.fhir.bulkexport.system.ChunkWriter">
+            <writer ref="com.ibm.fhir.jbatch.bulkdata.export.system.ChunkWriter">
                 <properties>
                     <property name="cos.api.key" value="#{jobParameters['cos.api.key']}"/>
                     <property name="cos.srvinst.id" value="#{jobParameters['cos.srvinst.id']}"/>
@@ -29,7 +29,7 @@
                     <property name="partition.resourcetype" value="#{partitionPlan['partition.resourcetype']}"/>
                 </properties>
             </writer>
-            <checkpoint-algorithm ref="com.ibm.fhir.bulkexport.common.CheckPointAlgorithm">
+            <checkpoint-algorithm ref="com.ibm.fhir.jbatch.bulkdata.export.common.CheckPointAlgorithm">
                 <properties>
                     <property name="cos.bucket.filemaxsize" value="#{jobParameters['cos.bucket.filemaxsize']}"/>
                     <property name="cos.bucket.filemaxresources" value="#{jobParameters['cos.bucket.filemaxresources']}"/>
@@ -37,13 +37,13 @@
             </checkpoint-algorithm>
         </chunk>
         <partition>
-            <mapper ref="com.ibm.fhir.bulkexport.patient.PatientExportPartitionMapper">
+            <mapper ref="com.ibm.fhir.jbatch.bulkdata.export.patient.PatientExportPartitionMapper">
                 <properties>
                     <property name="fhir.resourcetype" value="#{jobParameters['fhir.resourcetype']}"/>
                 </properties>
             </mapper>
-            <collector ref="com.ibm.fhir.bulkexport.system.ExportPartitionCollector"/>
-            <analyzer ref="com.ibm.fhir.bulkexport.system.ExportPartitionAnalyzer"/>
+            <collector ref="com.ibm.fhir.jbatch.bulkdata.export.system.ExportPartitionCollector"/>
+            <analyzer ref="com.ibm.fhir.jbatch.bulkdata.export.system.ExportPartitionAnalyzer"/>
         </partition>
     </step>
 </job>
