+++ fhir-bulkimportexport-webapp/src/main/java/META-INF/batch-jobs/FhirBulkImportChunkJob.xml
@@ -4,7 +4,7 @@
         <property name="fhir.cosreadsperdbbatch" value="#{jobParameters['fhir.cosreadsperdbbatch']}?:1;" />
     </properties>
     <listeners>
-        <listener ref="com.ibm.fhir.bulkimport.ImportJobListener">
+        <listener ref="com.ibm.fhir.jbatch.bulkdata.load.ImportJobListener">
             <properties>
                 <property name="fhir.dataSourcesInfo" value="#{jobParameters['fhir.dataSourcesInfo']}"/>
             </properties>
@@ -12,7 +12,7 @@
     </listeners>
     <step id="step1">
         <chunk checkpoint-policy="item" item-count="#{jobProperties['fhir.cosreadsperdbbatch']}">
-            <reader ref="com.ibm.fhir.bulkimport.ChunkReader">
+            <reader ref="com.ibm.fhir.jbatch.bulkdata.load.ChunkReader">
                 <properties>
                     <property name="import.partition.workitem" value="#{partitionPlan['import.partition.workitem']}"/>
                     <property name="partition.resourcetype" value="#{partitionPlan['partition.resourcetype']}"/>
@@ -25,7 +25,7 @@
                     <property name="import.fhir.storagetype" value="#{jobParameters['import.fhir.storagetype']}"/>
                 </properties>
             </reader>
-            <writer ref="com.ibm.fhir.bulkimport.ChunkWriter">
+            <writer ref="com.ibm.fhir.jbatch.bulkdata.load.ChunkWriter">
                 <properties >
                     <property name="fhir.tenant" value="#{jobParameters['fhir.tenant']}"/>
                     <property name="fhir.datastoreid" value="#{jobParameters['fhir.datastoreid']}"/>
@@ -41,7 +41,7 @@
             </writer>
         </chunk>
         <partition>
-            <mapper ref="com.ibm.fhir.bulkimport.ImportPartitionMapper">
+            <mapper ref="com.ibm.fhir.jbatch.bulkdata.load.ImportPartitionMapper">
                 <properties>
                     <property name="cos.api.key" value="#{jobParameters['cos.api.key']}"/>
                     <property name="cos.srvinst.id" value="#{jobParameters['cos.srvinst.id']}"/>
@@ -53,7 +53,7 @@
                     <property name="import.fhir.storagetype" value="#{jobParameters['import.fhir.storagetype']}"/>
                 </properties>
             </mapper>
-            <collector ref="com.ibm.fhir.bulkimport.ImportPartitionCollector">
+            <collector ref="com.ibm.fhir.jbatch.bulkdata.load.ImportPartitionCollector">
                 <properties>
                     <property name="cos.api.key" value="#{jobParameters['cos.api.key']}"/>
                     <property name="cos.srvinst.id" value="#{jobParameters['cos.srvinst.id']}"/>
@@ -63,7 +63,7 @@
                     <property name="cos.operationoutcomes.bucket.name" value="#{jobParameters['cos.operationoutcomes.bucket.name']}"/>
                 </properties>
             </collector>
-            <analyzer ref="com.ibm.fhir.bulkimport.ImportPartitionAnalyzer"/>
+            <analyzer ref="com.ibm.fhir.jbatch.bulkdata.load.ImportPartitionAnalyzer"/>
         </partition>
     </step>
 </job>
